requirements.txt
streamlit
requests
python-docx
pandas
streamlit-quill
google-auth-oauthlib
google-api-python-client